refactor(common): establish transaction fee domain owner
