+++ src/libs/portfolio-common/portfolio_common/domain/transaction/__init__.py
@@ -1,0 +1,1 @@
+"""Shared transaction domain policies consumed across Core boundaries."""
